YIPPEE MAP FINISHEDD

+++ Config/Settings.hpp
@@ -10,47 +10,47 @@ const float FRAME_RATE_LIMIT = 60.0f;
 const int SFX_VOLUME = 25;
 
 //For Boundaries 
-const int GRID1_WIDTH = 300;
-const int GRID1_HEIGHT = 300;
-const int GRID1_X = 490;
-const int GRID1_Y = 210;
+const float GRID1_WIDTH = 300.f;
+const float GRID1_HEIGHT = 300.f;
+const float GRID1_X = 490.f;
+const float GRID1_Y = 210.f;
 
-const int GRID2_WIDTH = 300;
-const int GRID2_HEIGHT = 400;
-const int GRID2_X = 490;
-const int GRID2_Y = 160;
+const float GRID2_WIDTH = 300.f;
+const float GRID2_HEIGHT = 400.f;
+const float GRID2_X = 490.f;
+const float GRID2_Y = 160.f;
 
-const int GRID3_WIDTH = 300;
-const int GRID3_HEIGHT = 500;
-const int GRID3_X = 490;
-const int GRID3_Y = 110;
+const float GRID3_WIDTH = 300.f;
+const float GRID3_HEIGHT = 500.f;
+const float GRID3_X = 490.f;
+const float GRID3_Y = 110.f;
 
-const int GRID4_WIDTH = 400;
-const int GRID4_HEIGHT = 300;
-const int GRID4_X = 440;
-const int GRID4_Y = 210;
+const float GRID4_WIDTH = 400.f;
+const float GRID4_HEIGHT = 300.f;
+const float GRID4_X = 440.f;
+const float GRID4_Y = 210.f;
 
-const int GRID5_WIDTH = 400;
-const int GRID5_HEIGHT = 400;
-const int GRID5_X = 440;
-const int GRID5_Y = 160;
+const float GRID5_WIDTH = 400.f;
+const float GRID5_HEIGHT = 400.f;
+const float GRID5_X = 440.f;
+const float GRID5_Y = 160.f;
 
-const int GRID6_WIDTH = 400;
-const int GRID6_HEIGHT = 500;
-const int GRID6_X = 440;
-const int GRID6_Y = 110;
+const float GRID6_WIDTH = 400.f;
+const float GRID6_HEIGHT = 500.f;
+const float GRID6_X = 440.f;
+const float GRID6_Y = 110.f;
 
-const int GRID7_WIDTH = 500;
-const int GRID7_HEIGHT = 300;
-const int GRID7_X = 390;
-const int GRID7_Y = 210;
+const float GRID7_WIDTH = 500.f;
+const float GRID7_HEIGHT = 300.f;
+const float GRID7_X = 390.f;
+const float GRID7_Y = 210.f;
 
-const int GRID8_WIDTH = 500;
-const int GRID8_HEIGHT = 400;
-const int GRID8_X = 390;
-const int GRID8_Y = 160;
+const float GRID8_WIDTH = 500.f;
+const float GRID8_HEIGHT = 400.f;
+const float GRID8_X = 390.f;
+const float GRID8_Y = 160.f;
 
-const int GRID9_WIDTH = 500;
-const int GRID9_HEIGHT = 500;
-const int GRID9_X = 390;
-const int GRID9_Y = 110;
+const float GRID9_WIDTH = 500.f;
+const float GRID9_HEIGHT = 500.f;
+const float GRID9_X = 390.f;
+const float GRID9_Y = 110.f;
